Add OFFSET clause to our grammar and update some of the scala diagram generation
Summary:
I've added the OFFSET clause to our grammar. In addition to this, made it simpler to run
the scala diagram_generator by making it a tool that can be run on the command line with arguments.
Apart from this updated the README.md with instructions on how to generate the diagrams and update
our docs repository.

Test Plan: Run it locally.

Reviewers: mihnea

Reviewed By: mihnea

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D5077