Remove restriction on specifying output folder
The commandline argument passed in for output folder gets checked for
containing the substring `syntax_resources`.  However, that is not
flexible such as when the working directory upon execution of the jar
file is the `syntax_resources` directory itself and the output folder is
given as `.`.  A workaround is to pass `../syntax_resources`; however,
that is cumbersome.  Remove this restriction on output folder by
resolving it to absolute path.