Fix help message; accept relative output folder (#1)
* Fix issues with help message

The help message printed from `printHelpAndExit` has some incorrect
information.  The generated diagrams are incorrectly associated with
`*.grammar.md`, and the example usage has inconsistent and incorrect
paths.  Fix both of these.

* Remove restriction on specifying output folder

The commandline argument passed in for output folder gets checked for
containing the substring `syntax_resources`.  However, that is not
flexible such as when the working directory upon execution of the jar
file is the `syntax_resources` directory itself and the output folder is
given as `.`.  A workaround is to pass `../syntax_resources`; however,
that is cumbersome.  Remove this restriction on output folder by
resolving it to absolute path.