Fix path building getGlobalRulePrefix
The current way to `getGlobalRulePrefix` uses `getAbsoluteFile`.  This
has problems when the given path consists of `.` and `..`.  Not only is
the generated prefix inconsistent but also the prefix may incorrectly
add too many `..`, resulting in broken links.  Switch over to using
`getCanonicalFile` to resolve both issues.  Also catch `IOException` for
`getCanonicalFile` with an error message.