Generate diagrams consistently (#2)
* Note the need for fonts in usage instructions

For consistent diagram generation, certain fonts need to be installed.
Add that instruction in the usage section of the README.

* Fix path building getGlobalRulePrefix

The current way to `getGlobalRulePrefix` uses `getAbsoluteFile`.  This
has problems when the given path consists of `.` and `..`.  Not only is
the generated prefix inconsistent but also the prefix may incorrectly
add too many `..`, resulting in broken links.  Switch over to using
`getCanonicalFile` to resolve both issues.  Also catch `IOException` for
`getCanonicalFile` with an error message.

* Enforce having font installed

Exit with error if the "Verdana" font cannot be found.  That way, font
availability is enforced, and there is less of a chance that diagrams
get generated differently.

* Move most font logic to RRDiagramToSVG

Move `isFontInstalled` and the font family name constant out of `Main`
to `RRDiagramToSVG`.  At `RRDiagramToSVG`, there are some hard-coded
`"Verdana"` font family names that should be replaced with the constant.
Then, that should eliminate all the hard-coding of `"Verdana"`.

* Fix README markdown multi-line numbered list

The README does not properly format its first numbered list.  Some items
of the numbered list span multiple lines; however, those lines aren't
indented, so they break the list apart.  This resets the numbering
counter, causing it to show 1 1 2 instead of 1 2 3.  Add indentation to
fix this.

* Generate files in sort order

Use `Arrays.sort` to generate files in an alphabetical order.  That way,
generation across different platforms is in the same order, making for
more consistent logging.

* Use binary search for finding font

The current scheme of looking for the existence of the right font does a
linear scan.  Use a built-in binary search algorithm instead, improving
runtime performance.

* Deduplicate usage instructions

The usage instructions in this README are also found on the Yugabyte
docs README.  To avoid having to maintain copies of the same README,
simply reference the usage instructions of the Yugabyte docs README.

* Bump version to 0.9.4-yb-1